+++ Assets/RaceGame/Scripts/Core/Interface/IGameSetting.cs
@@ -4,6 +4,6 @@
     {
         public bool StartFromTitle { get; set; }
         public PlayType PlayType { get; set; }
-        public int LocalPlayerID { get; set; }
+        public long LocalPlayerID { get; set; }
     }
 }
